+++ src/Section/TodayIs/TodayIs.js
@@ -1,0 +1,6 @@
+let today = new Date();
+function todayIs() {
+  return today.toLocaleDateString();
+}
+
+export { todayIs, today };
